Added Interrogation Room
Added an interrogation room for combat.

The floor consists of red tiles that will disappear a few seconds after they have been stepped on. The player will take considerable damage if they fall under the red tile floor; then the player will respawn back at the entrance of the room. The point is to encourage the player to parkour via jumping and using the arm abilities.

In the middle of the room is the actual interrogation cell, containing a security-camera turret that acts as sort of a mini boss (meaning it is stronger). I temporarily put the cafeteria table and chair there to decorate the cell, though they could be replaced with new models.