Read Description
It is f**king annoying that the player can't walk up the stairs properly. Now the player can. However, there is a bug that the player can't stop jumping when standing on stairs. I know how to fix this, but I can't figure out how to get the input system working.

The changes are made in movement.cs starting on line 188

My failed bugfix is annotated there.