moar of unfucking movement
You can now walk up slopes normally without hopping.

And also has anyone noticed the fact that you slide when walking on two bars (or planks or whatever they're meant to be)? Apparantly Raycast thinks that you are not on the ground and should therefore slide and not jump. I made a temporary fix to that issue, but of course it isnt permanent because I have no f**king idea how to make Physics.BoxCast work